* Many files - Undo more type casts to separate out the tricky changes

+++ lib/common/CollectInBufferStream.cpp
@@ -63,7 +63,7 @@ int CollectInBufferStream::Read(void *pBuffer, int NBytes, int Timeout)
 	// Adjust to number of bytes left
 	if(NBytes > (mBytesInBuffer - mReadPosition))
 	{
-		NBytes = (int)(mBytesInBuffer - mReadPosition);
+		NBytes = (mBytesInBuffer - mReadPosition);
 	}
 	ASSERT(NBytes >= 0);
 	if(NBytes <= 0) return 0;	// careful now
