+++ lib/common/MemBlockStream.cpp
@@ -113,7 +113,7 @@ int MemBlockStream::Read(void *pBuffer, int NBytes, int Timeout)
 	// Adjust to number of bytes left
 	if(NBytes > (mBytesInBuffer - mReadPosition))
 	{
-		NBytes = (int)(mBytesInBuffer - mReadPosition);
+		NBytes = (mBytesInBuffer - mReadPosition);
 	}
 	ASSERT(NBytes >= 0);
 	if(NBytes <= 0) return 0;	// careful now
